successfully create seeders

+++ models/customer.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, DataTypes) => {
       name: DataTypes.STRING,
       email: DataTypes.STRING,
       password: DataTypes.STRING,
-      phone_number: DataTypes.INTEGER,
+      phone_number: DataTypes.BIGINT,
       gender: DataTypes.ENUM("Laki-laki", "Perempuan"),
     },
     {
